Unflake `segment cache (CDN cache busting)` redirect test
[Flakiness Metrics](https://app.datadoghq.com/ci/test/runs?query=test_level%3Atest%20env%3Aci%20%40git.repository.id%3Agithub.com%2Fvercel%2Fnext.js%20%40test.service%3Anextjs%20%40test.status%3Afail%20%40test.name%3A%22segment%20cache%20%28CDN%20cache%20busting%29%20perform%20fully%20prefetched%20navigation%20when%20a%20third-party%20proxy%20performs%20a%20redirect%22%20%40git.branch%3Acanary&agg_m=count&agg_m_source=base&agg_t=count&currentTab=overview&eventStack=&fromUser=false&index=citest&start=1750359565225&end=1750964365225&paused=false)

When a prefetch is responded with a redirect by a third-party proxy, we
need to give the browser time to follow the redirect, before continuing
our request assertions inside of `act`. We do that by waiting for the
very next request to respond. Otherwise, we might exit early, before the
request was added to the pending requests set.